requirements.txt
Pillow==6.0.0
numpy==1.16.4
